+++ jabgui/src/main/java/org/jabref/gui/groups/GroupNodeViewModel.java
@@ -27,6 +27,7 @@ import org.jabref.gui.util.CustomLocalDragboard;
 import org.jabref.gui.util.DroppingMouseLocation;
 import org.jabref.gui.util.UiTaskExecutor;
 import org.jabref.logic.groups.GroupsFactory;
+import org.jabref.logic.l10n.Localization;
 import org.jabref.logic.layout.format.LatexToUnicodeFormatter;
 import org.jabref.logic.search.SearchContext;
 import org.jabref.logic.util.BackgroundTask;
@@ -437,12 +438,7 @@ public class GroupNodeViewModel {
     public boolean hasAllSuggestedGroups() {
         return hasSimilarSearchGroup(GroupsFactory.createWithoutFilesGroup())
                 && hasSimilarSearchGroup(GroupsFactory.createWithoutGroupsGroup())
-                && hasSimilarSearchGroup(GroupsFactory.createRankParentGroup())
-                && hasSimilarSearchGroup(GroupsFactory.createRelevanceParentGroup())
-                && hasSimilarSearchGroup(GroupsFactory.createQualityParentGroup())
-                && hasSimilarSearchGroup(GroupsFactory.createPrintedParentGroup())
-                && hasSimilarSearchGroup(GroupsFactory.createPriorityParentGroup())
-                && hasSimilarSearchGroup(GroupsFactory.createReadStatusParentGroup());
+                && getGroupNode().findGroupByName(Localization.lang("Marking and Grading")).isPresent();
     }
 
     public boolean canAddEntriesIn() {